refactor(webui): organize application shell
