+++ webui/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { ThemeProvider } from "next-themes";
 
-import { App } from "./App";
+import { App } from "@/app/App";
 import { Toaster } from "@/components/ui/sonner";
 import { LocaleProvider } from "@/i18n/locale";
 import { ThemeControllerProvider } from "@/themes/theme-controller";
