+++ webui/src/models/workspace.ts
@@ -1,0 +1,6 @@
+export const workspaces = ["overview", "events", "topology", "runtimes", "plugins", "configuration"] as const;
+export type Workspace = (typeof workspaces)[number];
+
+export function isWorkspace(value: string): value is Workspace {
+  return workspaces.includes(value as Workspace);
+}
